refactor: 전체조회수 엔티티명 변경 (#407)
