+++ src/main/java/com/festival/common/data/Data.java
@@ -1,23 +1,14 @@
 package com.festival.common.data;
 
-import com.festival.domain.booth.controller.dto.BulkInsertBooth;
 import com.festival.domain.booth.repository.BoothJdbcRepository;
-import com.festival.domain.image.model.Image;
 import com.festival.domain.image.repository.ImageRepository;
-import com.festival.domain.image.service.ImageService;
-import com.festival.domain.member.dto.MemberJoinReq;
 import com.festival.domain.member.service.MemberService;
-import com.festival.domain.viewcount.ViewCount;
-import com.festival.domain.viewcount.repository.ViewCountRepository;
-import org.hibernate.annotations.CurrentTimestamp;
+import com.festival.domain.viewcount.repository.HomeRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Profile({"dev"})
 @Configuration
@@ -30,7 +21,7 @@ public class Data {
             MemberService memberService,
             ImageRepository imageRepository,
             BoothJdbcRepository boothJdbcRepository,
-            ViewCountRepository viewCountRepository
+            HomeRepository homeRepository
     ) {
         return args -> {
