+++ src/main/java/com/festival/domain/viewcount/controller/ViewCountController.java
@@ -1,6 +1,6 @@
 package com.festival.domain.viewcount.controller;
 
-import com.festival.domain.viewcount.service.ViewCountService;
+import com.festival.domain.viewcount.service.HomeService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -12,11 +12,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping(value = "/api/v2/view-count")
 public class ViewCountController {
-    private final ViewCountService viewCountService;
+    private final HomeService homeService;
 
     @GetMapping
     public ResponseEntity<Void> increase(HttpServletRequest httpServletRequest){
-        viewCountService.increase(httpServletRequest.getRemoteAddr());
+        homeService.increase(httpServletRequest.getRemoteAddr());
         return ResponseEntity.ok().build();
     }
 }
