+++ src/main/java/com/festival/domain/viewcount/repository/HomeRepository.java
@@ -1,0 +1,7 @@
+package com.festival.domain.viewcount.repository;
+
+import com.festival.domain.viewcount.Home;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HomeRepository extends JpaRepository<Home, Long> {
+}
