--- src/main/java/com/festival/domain/viewcount/repository/ViewCountRepository.java
@@ -1,7 +1,0 @@
-package com.festival.domain.viewcount.repository;
-
-import com.festival.domain.viewcount.ViewCount;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ViewCountRepository extends JpaRepository<ViewCount, Long> {
-}
